Wrap the material with batched mesh uniforms
